requirements.txt
certifi==2022.5.18.1
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
certifi==2022.5.18.1
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
certifi==2022.5.18.1
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
certifi==2022.5.18.1
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
certifi==2022.5.18.1
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
certifi==2022.5.18.1
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
